requirements.txt
# from setup.py
cryptography~=2.0.2
requests[security]~=2.18.4
PyYAML~=3.11
python-dateutil~=2.6.1
pytz>=2017.2
six~=1.11.0
boto3~=1.4.4
